requirements.txt
fpdf
numpy
biopython
streamlit
